IRO_platform/requirements.txt
# requirements.txt
Django==4.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.2.7
dj-database-url==0.5.0
django-tenants==3.5.0
django-crispy-forms==2.3
djangorestframework==3.14
crispy-bootstrap5==0.7
django-guardian==2.4.0
whitenoise==6.4.0
